requirements.txt
cycler==0.10.0
decorator==4.4.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
pyparsing==2.4.4
pystache==0.5.4
python-dateutil==2.8.1
PyYAML==5.1.2
six==1.13.0
